pyproject.toml
[project]
name = "cadastro-pontos"
version = "0.1.0"
description = ""
authors = [
    {name = "Alessandro"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "pandas (>=2.3.1,<3.0.0)",
    "numpy (>=2.3.1,<3.0.0)",
    "plotly-express (>=0.4.1,<0.5.0)",
    "folium (>=0.20.0,<0.21.0)",
    "streamlit (>=1.46.1,<2.0.0)",
    "geopandas (>=1.0.0,<=1.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "streamlit-folium (>=0.25.0,<0.26.0)",
    "docker (>=7.1.0,<8.0.0)",
    "supabase (>=2.16.0,<3.0.0)",
    "tinydb (>=4.8.2,<5.0.0)",
    "geobr (>=0.2.2,<0.3.0)",
    "psycopg2 (>=2.9.10,<3.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
